# front/openai/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "JdR – Terres du Milieu", "version": "0.1.0" },
  "paths": {
    "/api/characters/": {
      "get": {
        "summary": "List Characters",
        "description": "Récupère la liste de tous les personnages disponibles.\n\nReturns:\n    CharacterList: Une liste de personnages disponibles.",
        "operationId": "list_characters_api_characters__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CharacterList" }
              }
            }
          }
        }
      }
    },
    "/api/combat/attack": {
      "post": {
        "summary": "Attack Endpoint",
        "description": "Endpoint pour effectuer une attaque dans le nouveau système de combat.\n\nParameters:\n- attacker_id (str): ID de l'attaquant.\n- target_id (str): ID de la cible.\n- attack_value (int): Valeur d'attaque.\n- combat_state (dict): Etat courant du combat (sérialisé).\n\nReturns:\n- dict: Etat du combat mis à jour.",
        "operationId": "attack_endpoint_api_combat_attack_post",
        "parameters": [
          {
            "name": "attacker_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Attacker Id" }
          },
          {
            "name": "target_id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Target Id" }
          },
          {
            "name": "attack_value",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "title": "Attack Value" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Combat State"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/": {
      "get": {
        "summary": "List Scenarios",
        "description": "### list_scenarios\n**Description :** Endpoint pour lister les scénarios disponibles et en cours.\n**Paramètres :** Aucun.\n**Retour :** `ScenarioList` contenant une liste de scénarios avec leurs états.",
        "operationId": "list_scenarios_api_scenarios__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ScenarioList" }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenario_file}": {
      "get": {
        "summary": "Get Scenario Details",
        "description": "### get_scenario_details\n**Description :** Récupère le contenu d'un scénario (fichier Markdown) à partir de son nom de fichier.\n**Paramètres :**\n- `scenario_file` (str) : Le nom du fichier du scénario (ex: Les_Pierres_du_Passe.md).\n**Retour :** Le contenu du fichier Markdown du scénario, ou une erreur 404 si le fichier n'existe pas.",
        "operationId": "get_scenario_details_api_scenarios__scenario_file__get",
        "parameters": [
          {
            "name": "scenario_file",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Scenario File" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/start": {
      "post": {
        "summary": "Start Scenario",
        "description": "### start_scenario\n**Description :** Démarre un scénario avec un personnage spécifique, retourne l'id de session et déclenche le début du scénario avec le LLM.\n**Paramètres :**\n- `request` (StartScenarioRequest) : Contient le nom du scénario et l'identifiant du personnage.\n**Retour :** Un objet JSON avec session_id, scenario_name, character_id, message et la réponse initiale du LLM.\n**Exceptions :**\n- HTTPException 409 : Si une session existe déjà avec le même scénario et personnage.",
        "operationId": "start_scenario_api_scenarios_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StartScenarioRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/play": {
      "post": {
        "summary": "Play Scenario",
        "description": "### play_scenario\n**Description :** Envoie un message au MJ (LLM) pour jouer le scénario. Récupère automatiquement les informations de session (personnage et scénario).\n**Paramètres :**\n- `session_id` (UUID) : Identifiant de la session de jeu (query parameter).\n- `request` (PlayScenarioRequest) : Contient le message du joueur.\n**Retour :**\n- `response` (str) : Réponse du Maître du Jeu\n- `tool_calls` (list) : Liste des appels d'outils effectués (si applicable)",
        "operationId": "play_scenario_api_scenarios_play_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlayScenarioRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/history/{session_id}": {
      "get": {
        "summary": "Get Scenario History",
        "description": "### get_scenario_history\n**Description :** Récupère l'historique complet des messages de la session de jeu spécifiée.\n**Paramètres :**\n- `session_id` (UUID) : Identifiant de la session de jeu.\n**Retour :**\n- `history` (list) : Liste ordonnée de tous les messages (user, assistant, tool, etc.) de la session.",
        "operationId": "get_scenario_history_api_scenarios_history__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Character": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "race": { "type": "string", "title": "Race" },
          "culture": { "type": "string", "title": "Culture" },
          "profession": { "type": "string", "title": "Profession" },
          "caracteristiques": {
            "additionalProperties": { "type": "integer" },
            "type": "object",
            "title": "Caracteristiques"
          },
          "competences": {
            "additionalProperties": { "type": "integer" },
            "type": "object",
            "title": "Competences"
          },
          "hp": { "type": "integer", "title": "Hp", "default": 100 },
          "inventory": {
            "items": { "$ref": "#/components/schemas/Item" },
            "type": "array",
            "title": "Inventory",
            "default": []
          },
          "equipment": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Equipment",
            "default": []
          },
          "spells": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Spells",
            "default": []
          },
          "equipment_summary": {
            "anyOf": [
              {
                "additionalProperties": { "type": "number" },
                "type": "object"
              },
              { "type": "null" }
            ],
            "title": "Equipment Summary"
          },
          "culture_bonuses": {
            "anyOf": [
              {
                "additionalProperties": { "type": "integer" },
                "type": "object"
              },
              { "type": "null" }
            ],
            "title": "Culture Bonuses"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "race",
          "culture",
          "profession",
          "caracteristiques",
          "competences"
        ],
        "title": "Character"
      },
      "CharacterList": {
        "properties": {
          "characters": {
            "items": { "$ref": "#/components/schemas/Character" },
            "type": "array",
            "title": "Characters"
          }
        },
        "type": "object",
        "required": ["characters"],
        "title": "CharacterList"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Item": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "weight": { "type": "number", "title": "Weight" },
          "base_value": { "type": "number", "title": "Base Value" }
        },
        "type": "object",
        "required": ["id", "name", "weight", "base_value"],
        "title": "Item"
      },
      "PlayScenarioRequest": {
        "properties": { "message": { "type": "string", "title": "Message" } },
        "type": "object",
        "required": ["message"],
        "title": "PlayScenarioRequest"
      },
      "ScenarioList": {
        "properties": {
          "scenarios": {
            "items": { "$ref": "#/components/schemas/ScenarioStatus" },
            "type": "array",
            "title": "Scenarios"
          }
        },
        "type": "object",
        "required": ["scenarios"],
        "title": "ScenarioList"
      },
      "ScenarioStatus": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "status": { "type": "string", "title": "Status" },
          "session_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Session Id"
          },
          "scenario_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Scenario Name"
          },
          "character_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Character Name"
          }
        },
        "type": "object",
        "required": ["name", "status"],
        "title": "ScenarioStatus"
      },
      "StartScenarioRequest": {
        "properties": {
          "scenario_name": { "type": "string", "title": "Scenario Name" },
          "character_id": { "type": "string", "title": "Character Id" }
        },
        "type": "object",
        "required": ["scenario_name", "character_id"],
        "title": "StartScenarioRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}